feat: Set AAPT2 binary in Aapt2Jni
