+++ build-tools/build-tests/src/test/resources/TestProject/app/.gitignore
@@ -1,0 +1,1 @@
+/build
